requirements.txt
Flask==2.3.3
python-decouple==3.8
mysql-connector-python==8.0.33
pymysql==1.0.3